backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.3.2

# LLM
httpx==0.28.1

# File parsing
pymupdf==1.25.3
python-docx==1.2.0
python-multipart==0.0.20

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.1

# Utilities
python-dateutil==2.9.0
uuid6==2024.7.10
